Hot fix because Flask project somehow failed
