--- app_python/app/__init__.py
@@ -1,10 +1,0 @@
-from flask import Flask
-from app_python.app import routes
-
-
-def create_app():
-    app = Flask(__name__)
-    return app
-
-
-app = create_app()
